Did assignment 16

+++ 16_subseq/maxSeq.c
@@ -1,10 +1,10 @@
 #include<stdio.h>
 #include<stdlib.h>
 
-size_t maxSeq(size_t * array, size_t n)
+size_t maxSeq(int * array, size_t n)
 {size_t count=1;
   size_t temp=0;
-  size_t i;
+  int i;
   if(n==0)
     {return 0;
     }
@@ -14,10 +14,12 @@ size_t maxSeq(size_t * array, size_t n)
   else
     { for(i=0;i<n;i++)
 	{  if(array[i+1]>array[i])
-	  {count++;
+	  {
+	    count++;
 	  }
      else if(array[i+1]<=array[i])
-       { if(count>temp)
+       {
+	 if(count>temp)
 	 {temp=count;
 	  count=1;
 	}
@@ -28,6 +30,5 @@ size_t maxSeq(size_t * array, size_t n)
 	{ return temp;
     }
       else return count;
-      
-}
+      }
 }
